Moved RecordView into webapp/components/survey/Record
