+++ webapp/views/App/views/Data/Data.js
@@ -8,7 +8,7 @@ import { useSurveyInfo } from '@webapp/store/survey'
 
 import ModuleSwitch from '@webapp/components/moduleSwitch'
 import SurveyDefsLoader from '@webapp/components/survey/SurveyDefsLoader'
-import RecordView from '@webapp/loggedin/surveyViews/record/recordView'
+import { RecordView } from '@webapp/components/survey/Record'
 
 import { resetDataVis } from '@webapp/views/App/views/Data/Explorer/actions'
