+++ webapp/views/App/views/Designer/FormDesigner/FormDesigner.js
@@ -1,12 +1,11 @@
 import React from 'react'
 import { useSelector } from 'react-redux'
 
-import SurveyFormView from '@webapp/loggedin/surveyViews/surveyForm/surveyFormView'
-import RecordView from '@webapp/loggedin/surveyViews/record/recordView'
-
 import { RecordState } from '@webapp/store/ui/record'
-
 import { useAuthCanEditSurvey } from '@webapp/store/user'
+
+import SurveyFormView from '@webapp/loggedin/surveyViews/surveyForm/surveyFormView'
+import { RecordView } from '@webapp/components/survey/Record'
 
 const FormDesigner = () => {
   const canEditDef = useAuthCanEditSurvey()
